requirement.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
langchain_groq
langchain_core
langchain
langchain_community
ipykernel
pymupdf